Remove gulp; Use Makefile-based build system
Now with 1970s tech to solve 2021 problems.

This breaks building on Windows; use WSL instead or run the commands
manually.

Use `make build` for a development build; `make watch` to watch for
changes and start a development server; `make run` to build and start
the development server.

This now also enforces some style rules via eslint (auto-fix on build).
Note that the coding style is subject to change; and this is a temporary
solution until we settled on a style guide.

fixes #2